+++ src/solidity/contracts/emitter/EventEmitterCaller.sol
@@ -1,8 +1,8 @@
 pragma solidity ^0.8.0;
 
 contract EventEmitterCaller {
-    event CallerSimpleEvent(uint indexed id, string message, address indexed sender);
-    event CallerAddressEvent(uint indexed id, address indexed _address);
+    event CallerSimpleEvent(uint256 indexed id, string message, address indexed sender);
+    event CallerAddressEvent(uint256 indexed id, address indexed _address);
 
     address private emitterAddress;
 
@@ -14,7 +14,7 @@ contract EventEmitterCaller {
         return string(abi.encodePacked(_prefix, _base));
     }
 
-    function callEmitSimpleEvent(uint id, string memory _prepend, string memory _message) public {
+    function callEmitSimpleEvent(uint256 id, string memory _prepend, string memory _message) public {
         (bool success, ) = emitterAddress.call(
             abi.encodeWithSignature("emitSimpleEvent(uint256,string)", id, _message)
         );
@@ -22,7 +22,7 @@ contract EventEmitterCaller {
         emit CallerSimpleEvent(id, prepend(_prepend, _message), msg.sender);
     }
 
-    function callEmitAddressEvent(uint id, address _address) public {
+    function callEmitAddressEvent(uint256 id, address _address) public {
         (bool success, ) = emitterAddress.call(
             abi.encodeWithSignature("emitAddressEvent(uint256,address)", id, _address)
         );
